+++ src/test/kotlin/com/github/victorrentea/livecoding/InspectionParameterizedTestBase.kt
@@ -24,6 +24,7 @@ abstract class InspectionParameterizedTestBase(val fileName: String) : LightJava
     @Test
     fun test() {
         myFixture.enableInspections(inspectionClass())
+//        PsiTestUtil.addLibrary(, "C:\\Users\\victo\\.m2\\repository\\org\\projectlombok\\lombok\\1.18.22\\lombok-1.18.22.jar")
 
         myFixture.configureByFiles(fileName);