+++ src/test/live-coding-playground/src/main/java/playground/DataOnEntity.java
@@ -1,18 +1,11 @@
 package playground;
 
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
 import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import java.util.List;
 
 
-@ToString
-@Setter
-@Getter
 @Entity
 public class DataOnEntity {