+++ src/test/live-coding-playground/src/test/java/lombok/AddRAC.java
@@ -1,6 +1,5 @@
 package lombok;
 
-@RequiredArgsConstructor
 public class AddRAC {
     private final int x;
 }