+++ src/test/live-coding-playground/src/test/java/lombok/replacerac/ReplaceRAC_OneArg.java
@@ -1,8 +1,5 @@
 package lombok.replacerac;
 
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
 public class ReplaceRAC_OneArg {
     private final int dummy;