+++ src/test/live-coding-playground/src/test/java/lombok/replacerac/ReplaceRAC_TwoArg.java
@@ -1,8 +1,5 @@
 package lombok.replacerac;
 
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
 public class ReplaceRAC_TwoArg {
     private final int dummy;
     private final int dummy2;