+++ src/test/live-coding-playground/src/test/java/lombok/replacerac/ReplaceRAC_TwoArg_Component.java
@@ -1,9 +1,7 @@
 package lombok.replacerac;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
-@RequiredArgsConstructor
 @Component
 public class ReplaceRAC_TwoArg_Component {
     private final int dummy;