+++ src/test/live-coding-playground/src/test/java/lombok/replacerac/ReplaceRAC_TwoArg_Service.java
@@ -1,9 +1,7 @@
 package lombok.replacerac;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-@RequiredArgsConstructor
 @Service
 public class ReplaceRAC_TwoArg_Service {
     private final int dummy;