+++ CHANGELOG.md
@@ -37,3 +37,5 @@
 - tech re-release
 ## 0.0.10
 - Added inspection to remove final from parameters and local variables (no highlight, only fix)
+- Added nice balloon after static import completes
+-