+++ CHANGELOG.md
@@ -38,4 +38,4 @@
 ## 0.0.10
 - Added inspection to remove final from parameters and local variables (no highlight, only fix)
 - Added nice balloon after static import completes
-- 
+- Added coloring the test editors in green (beta)